% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_analyze.R
\name{mr_analyze}
\alias{mr_analyze}
\title{Title}
\usage{
mr_analyze(exposure, outcome, p)
}
\arguments{
\item{exposure}{A summary-statistic GWAS dataset to be used as the exposure}

\item{outcome}{A summary-statistic GWAS dataset to be used as the outcome}

\item{p}{p-value threshold for SNP retrieval from datasets}
}
\value{
A chart of MR results from various methods (an S4 object)
}
\description{
Title
}
